refactor(core): Move ua_timer.* to /arch/common
