+++ src/client/ua_client_internal.h
@@ -20,7 +20,7 @@
 
 #include "open62541_queue.h"
 #include "ua_securechannel.h"
-#include "ua_timer.h"
+#include "common/ua_timer.h"
 
 _UA_BEGIN_DECLS
