+++ src/server/ua_subscription.h
@@ -22,7 +22,7 @@
 #include <open62541/plugin/nodestore.h>
 
 #include "ua_session.h"
-#include "ua_timer.h"
+#include "common/ua_timer.h"
 #include "ua_util_internal.h"
 
 _UA_BEGIN_DECLS
