+++ src/ua_securechannel.h
@@ -18,6 +18,7 @@
 #include <open62541/transport_generated.h>
 
 #include "open62541_queue.h"
+#include "ua_util_internal.h"
 #include "ua_connection_internal.h"
 
 _UA_BEGIN_DECLS
