+++ tests/CMakeLists.txt
@@ -23,6 +23,7 @@ include_directories(${open62541_BUILD_INCLUDE_DIRS})
 # ua_server_internal.h
 include_directories("${PROJECT_SOURCE_DIR}/src")
 include_directories("${PROJECT_SOURCE_DIR}/src/server")
+include_directories("${PROJECT_SOURCE_DIR}/arch/common")
 # testing_clock.h
 include_directories("${CMAKE_CURRENT_SOURCE_DIR}/testing-plugins")
 # #include <src_generated/<...>.h>
